Week_06/app/website_loader.py
import requests
from bs4 import BeautifulSoup

def load_website(url: str) -> str:
    response = requests.get(url, timeout=10)
    soup = BeautifulSoup(response.text, "html.parser")

    for script in soup(["script", "style"]):
        script.extract()

    text = soup.get_text(separator="\n")
    return text
