Week_06/requirements.txt
fastapi
uvicorn
langchain==0.1.17
langchain-openai==0.0.8
langchain-community==0.0.34
langchain-text-splitters==0.0.1
chromadb
pypdf
python-docx
tiktoken
python-multipart
beautifulsoup4
requests